[FIX] ErrorType, SucessType 이름 명확하게 변경
